refactor: `OpeningDate` Domain 모델로 이동
